docker/requirements.txt
numpy == 1.22
flask == 2.1.1
pillow == 9.1.0
gunicorn == 20.1.0
pyyaml